+++ bin/common.ml
@@ -319,6 +319,7 @@ type t = {
   output : string option;
   repo : string option;
   conf : Conf.t;
+  check_time : Okra.Time.t option;
 }
 
 let setup () =
@@ -358,7 +359,9 @@ let term =
   and+ conf = conf
   and+ calendar = calendar
   and+ output = output
-  and+ repo = repo in
+  and+ repo = repo
+  and+ engineer_report = engineer_report in
+  let check_time = if engineer_report then Some (Okra.Time.days 5.) else None in
   {
     okr_db;
     filter;
@@ -369,6 +372,7 @@ let term =
     output;
     repo;
     okr_db_state = None;
+    check_time;
   }
 
 let okr_db t =
@@ -434,6 +438,7 @@ let with_names t = t.printconf.names
 let with_links t = t.printconf.links
 let with_description t = t.printconf.descriptions
 let conf t = t.conf
+let check_time t = t.check_time
 
 let output ?(input_files = []) ?(in_place = false) t =
   match t.output with