+++ bin/lint.ml
@@ -31,6 +31,7 @@ let run conf =
   let collect_errors name ic =
     match
       Okra.Lint.lint ?okr_db:(Common.okr_db conf.c)
+        ?check_time:(Common.check_time conf.c)
         ~include_sections:(Common.include_sections conf.c)
         ~ignore_sections:(Common.ignore_sections conf.c)
         ic