+++ lib/lint.mli
@@ -19,6 +19,7 @@ type lint_error =
   | Format_error of (int * string) list
   | No_time_found of int option * string
   | Invalid_time of { lnum : int option; title : string; entry : string }
+  | Invalid_total_time of string * Time.t
   | Multiple_time_entries of int option * string
   | No_work_found of int option * string
   | No_KR_ID_found of int option * string
@@ -32,6 +33,7 @@ val lint :
   ?okr_db:Masterdb.t ->
   ?include_sections:string list ->
   ?ignore_sections:string list ->
+  ?check_time:Time.t ->
   in_channel ->
   lint_result
 
@@ -39,6 +41,7 @@ val lint_string_list :
   ?okr_db:Masterdb.t ->
   ?include_sections:string list ->
   ?ignore_sections:string list ->
+  ?check_time:Time.t ->
   string list ->
   lint_result
 (** [lint_string_list] is like {!lint} except the input is a list of lines *)