+++ lib/team.ml
@@ -46,8 +46,11 @@ let lint_member_week admin_dir member ~week ~year =
     | false -> Not_found
     | true -> (
         let ic = open_in fname in
+        (* We lint week-by-week so we use the default options of the [engineer]
+           mode. *)
         match
           Lint.lint ~include_sections:[ "Last week" ] ~ignore_sections:[] ic
+            ~check_time:(Time.days 5.)
         with
         | Ok () -> Complete
         | Error e -> Not_lint e)