+++ CHANGES.md
@@ -1,7 +1,8 @@
 ## unreleased
 
-### Changes
+### Changed
 
+- Lint: check that the total of days reported for each engineer is 5 days (#178, @gpetiot)
 - No special treatment for "OKR Updates" sections in reports (#211, @gpetiot)
 - Lookup okr-db in the repo directory (set by `--repo-dir`/`-C`) if `--okr-db` is not set (#210, @gpetiot)
 - Make github handles clickable in repo reports (#193, #207, @gpetiot)